Refactored util class to be under commons
