+++ src/main/java/seedu/address/model/person/timetable/TimetableDay.java
@@ -1,6 +1,6 @@
 package seedu.address.model.person.timetable;
 
-import seedu.address.logic.parser.TimetableParserUtil;
+import seedu.address.commons.util.TimetableParserUtil;
 
 /**
  * Represents a single day in a timetable
